refactor(llm): extract ChannelSectionFilter into its own header
Behavior-preserving PoC for decoupling the 5.5k-line LLM.cpp God object:
move the fully self-contained ChannelSectionFilter (hidden-channel thought
stripper, depends only on std::string, touches no Impl state) verbatim into
src/runner/ChannelSectionFilter.hpp and #include it. No logic change; both
call sites (strip_hidden_channel_sections, Impl streaming) resolve unchanged.

Verified: header compiles standalone (-std=c++17 -fsyntax-only); brace count
in LLM.cpp unchanged-balanced. Full build + golden run still required on an
AXCL/SDK box (not available on this workstation).
